+++ apps/ced-browser-fe/src/core/auth/authSelectors.ts
@@ -6,9 +6,9 @@ const EXPIRE_TIME = 30 * 60 * 1000; // 30 minuti in ms
 
 export const selectToken = (state: RootState) => state.auth.token;
 
-export const selectSavedAt = ({ auth }: RootState) => auth?.savedAt;
+const selectSavedAt = ({ auth }: RootState) => auth?.savedAt;
 
-export const selectRedirectToken = ({ auth }: RootState) => auth?.redirectToken;
+const selectRedirectToken = ({ auth }: RootState) => auth?.redirectToken;
 
 export const selectDeviceId = ({ auth }: RootState) => auth?.deviceId;