+++ apps/ced-card-request-fe/src/components/Checkbox/index.tsx
@@ -1,5 +1,0 @@
-import { Checkbox, type CheckboxProps } from '@mui/material';
-
-export const AppCheckbox = ({ sx, ...props }: CheckboxProps) => (
-  <Checkbox sx={sx} {...props} />
-);