+++ apps/ced-card-request-be/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ced-card-request-be
 
+## 0.2.5
+
+### Patch Changes
+
+- 8936451: add anti corruption layer to inps outbound adapter
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [1dfaa7e]
+  - @pagopa/io-core-adapter-tracing@0.0.7
+
 ## 0.2.3
 
 ### Patch Changes